refactor(display): extract setup-mode screen into its own module
Move DrawQR/DisplaySetupScreen out of main.cpp into src/setup_screen.cpp
and include/setup_screen.h, mirroring how display.cpp is already split
out. This lets the screen be compiled into the WASM simulator alongside
the other rendering code.
